ai_video_upscaler/requirements.txt
# Core dependencies
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=8.3.0
tqdm>=4.62.0

# PyTorch (will be installed separately for macOS)
# torch>=1.9.0
# torchvision>=0.10.0

# AI upscaling models
basicsr>=1.4.2
facexlib>=0.3.0
gfpgan>=1.3.5
realesrgan>=0.3.0 